+++ app/components/Snaps/SnapAvatar/SnapAvatar.styles.ts
@@ -1,6 +1,6 @@
 ///: BEGIN:ONLY_INCLUDE_IF(preinstalled-snaps,external-snaps)
 import { StyleSheet } from 'react-native';
-import { Theme } from '../../../../util/theme/models';
+import { Theme } from '../../../util/theme/models';
 
 /**
  *
