+++ app/components/Snaps/SnapAvatar/SnapAvatar.test.tsx
